requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain==0.3.0
langchain-community==0.3.0
langchain-text-splitters==0.3.0
langchain-huggingface==0.1.0
langchain-google-genai==2.0.1
langchain-core==0.3.0
chromadb
google-generativeai
pypdf
requests
sentence-transformers
python-multipart
sqlalchemy
pymysql